refactor: - Remove 3d dependencies - AvatarDownloader and AvatarManager as abstract class (for extensions)
