--- unity/Assets/3DAssets.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d1b68b513939a4766b1b6be7381e3333
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
