--- unity/Assets/3DAssets/BaseAnimationController.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7cf492234580245d399fe5a70ac83c43
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
