--- unity/Assets/3DAssets/Tpose.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2e097843fb0d2407cb6acbb9e636eaee
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
