--- unity/Assets/Prefabs/MYTY3DAvatarObjects.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: f30a0670d5fc14194b3bf9ede8a9e389
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
