--- unity/Assets/Scenes/3DAvatar.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: b24baef25963741d48be80fdec96a244
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
