--- unity/Assets/Scripts/Avatar/Avatar3DLoader.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: decc34eb30faa4d3b8c752dc533ce999
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
