--- unity/Assets/Scripts/Avatar/AvatarLoader.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f02204f6fd5f740ce8936bbce29dd407
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
