--- unity/Assets/Scripts/Avatar/AvatarManager.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 092253a6a0514d998d991e38e47f253a
-timeCreated: 1679989108
