+++ unity/Assets/Scripts/Avatar/Impl/AvatarDownloader2D.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 37cc105622af4b4dabfceb5c7c64eaaa
+timeCreated: 1682400193
