+++ unity/Assets/Scripts/Avatar/Impl/AvatarManager2D.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: adcbe84d31424bdba0050c5521965ba7
+timeCreated: 1682399308
