+++ unity/Assets/Scripts/Avatar/Interface.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: c7dbcd028f4d418fa1944fe2a9d861a2
+timeCreated: 1682397764
