+++ unity/Assets/Scripts/Avatar/Interface/AvatarManager.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 06717e740c89446f9fa814c9360497b7
+timeCreated: 1682398727
