+++ unity/Packages/manifest.json
@@ -1,6 +1,5 @@
 {
   "dependencies": {
-    "com.offlive.myty.myty-extension.3d": "https://github.com/Off-Live/myty-3d-avatar-extension.git?path=/Assets",
     "com.offlive.myty.myty-kit": "https://github.com/Off-Live/myty-kit.git?path=/Assets#pre-release/sdk-core-dev",
     "com.unity.2d.animation": "7.0.10",
     "com.unity.collab-proxy": "1.15.16",
@@ -13,9 +12,6 @@
     "com.unity.timeline": "1.6.4",
     "com.unity.ugui": "1.0.0",
     "com.unity.visualscripting": "1.7.6",
-    "com.vrmc.gltf": "https://github.com/vrm-c/UniVRM.git?path=/Assets/UniGLTF#v0.109.0",
-    "com.vrmc.vrm": "https://github.com/vrm-c/UniVRM.git?path=/Assets/VRM10#v0.109.0",
-    "com.vrmc.vrmshaders": "https://github.com/vrm-c/UniVRM.git?path=/Assets/VRMShaders#v0.109.0",
     "com.unity.modules.ai": "1.0.0",
     "com.unity.modules.androidjni": "1.0.0",
     "com.unity.modules.animation": "1.0.0",
