+++ unity/Packages/packages-lock.json
@@ -1,12 +1,5 @@
 {
   "dependencies": {
-    "com.offlive.myty.myty-extension.3d": {
-      "version": "https://github.com/Off-Live/myty-3d-avatar-extension.git?path=/Assets",
-      "depth": 0,
-      "source": "git",
-      "dependencies": {},
-      "hash": "cf9ef8afe3dffaf9749148ce4cfa5b6505e71b2f"
-    },
     "com.offlive.myty.myty-kit": {
       "version": "https://github.com/Off-Live/myty-kit.git?path=/Assets#pre-release/sdk-core-dev",
       "depth": 0,
@@ -216,35 +209,6 @@
         "com.unity.modules.jsonserialize": "1.0.0"
       },
       "url": "https://packages.unity.com"
-    },
-    "com.vrmc.gltf": {
-      "version": "https://github.com/vrm-c/UniVRM.git?path=/Assets/UniGLTF#v0.109.0",
-      "depth": 0,
-      "source": "git",
-      "dependencies": {
-        "com.vrmc.vrmshaders": "0.109.0",
-        "com.unity.modules.animation": "1.0.0"
-      },
-      "hash": "7affa42bcab90a1c673d50670d5642d6d7c29c29"
-    },
-    "com.vrmc.vrm": {
-      "version": "https://github.com/vrm-c/UniVRM.git?path=/Assets/VRM10#v0.109.0",
-      "depth": 0,
-      "source": "git",
-      "dependencies": {
-        "com.vrmc.vrmshaders": "0.109.0",
-        "com.vrmc.gltf": "0.109.0"
-      },
-      "hash": "7affa42bcab90a1c673d50670d5642d6d7c29c29"
-    },
-    "com.vrmc.vrmshaders": {
-      "version": "https://github.com/vrm-c/UniVRM.git?path=/Assets/VRMShaders#v0.109.0",
-      "depth": 0,
-      "source": "git",
-      "dependencies": {
-        "com.unity.modules.imageconversion": "1.0.0"
-      },
-      "hash": "7affa42bcab90a1c673d50670d5642d6d7c29c29"
     },
     "com.unity.modules.ai": {
       "version": "1.0.0",
